Inserto la multiplicación al trabajo final

+++ src/Calculator.java
@@ -1,5 +1,5 @@
 public class Calculator {
-    public float add(float num1, float num2){
+    public double add(double num1, double num2){
         return num1 + num2;
     }
     public double multiply(double num1, double num2){
